+++ .idea/copyright/copyright_AGPL_amatsukaze.xml
@@ -1,0 +1,7 @@
+<component name="CopyrightManager">
+  <copyright>
+    <option name="allowReplaceRegexp" value="2020" />
+    <option name="notice" value="Copyright (c) 2020-&amp;#36;today.year. kokoro-aya. All right reserved.&#10;Amatsukaze - A Playground Server implemented with ANTLR or Kotlin DSL&#10;&#10;This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or any later version.&#10;This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.&#10;You should have received a copy of the GNU Affero General Public License along with this program. If not, see &lt;https://www.gnu.org/licenses/&gt;.&#10;" />
+    <option name="myName" value="copyright-AGPL-amatsukaze" />
+  </copyright>
+</component>
