+++ .idea/copyright/profiles_settings.xml
@@ -1,0 +1,7 @@
+<component name="CopyrightManager">
+  <settings default="copyright-AGPL-amatsukaze">
+    <module2copyright>
+      <element module="Project Files" copyright="copyright-AGPL-amatsukaze" />
+    </module2copyright>
+  </settings>
+</component>
