+++ README.md
@@ -1,9 +1,15 @@
 # amatsukaze
-A compiler of a home made DSL integrated within a Playground
+~~A compiler of a home made DSL integrated within a Playground~~
 
-~~**Please be aware that this project is discontinued has moved to [here](https://github.com/kokoro-aya/shizuku). The new project is subject to be refactored by introducing a better DSL implementation and a more completed playground data structure. This repo will therefore be archived and few changes will be committed unless in case or urgency.**~~
+A Playground server implemented with ANTLR and/or Kotlin DSL.
 
-The project is currently under refactoring using the Kotlin DSL in lieu of homemade DSL using ANTLR. It's therefore uncertain whether the development will continue in this repo or will switch to Shizuku project.
+In the final release you will be able to choose from using my home made DSL or using Kotlin DSL.
+
+~~**Please be aware that this project is discontinued has moved to [here](https://github.com/kokoro-aya/shizuku). 
+The new project is subject to be refactored by introducing a better DSL implementation and a more completed playground
+data structure. This repo will therefore be archived and few changes will be committed unless in case or urgency.**~~
+
+The project is currently under refactoring using the Kotlin DSL in lieu of homemade DSL using ANTLR.
 
 For more information please review [the Wiki](https://github.com/kokoro-aya/amatsukaze/wiki) of this repo.
