--- src/main/kotlin/org/ironica/amatsukaze/PlaygroundFunctions.kt
@@ -1,4 +1,0 @@
-package org.ironica.amatsukaze
-
-enum class PF { Method, Property }
-enum class PFType { Self, Int, Color, None, Bool }
