--- src/main/kotlin/org/ironica/amatsukaze/Test2.kt
@@ -1,11 +1,0 @@
-package org.ironica.amatsukaze
-
-open class ABC
-
-class DEF: ABC()
-
-fun main() {
-    val a = DEF()
-    val b = listOf(a)
-    b.filterIsInstance<ABC>()
-}
