+++ src/main/kotlin/org/ironica/amatsukaze/playground/Playground.kt
@@ -1,9 +1,19 @@
-package org.ironica.amatsukaze
+/*
+ * Copyright (c) 2020-2021. kokoro-aya. All right reserved.
+ * Amatsukaze - A Playground Server implemented with ANTLR or Kotlin DSL
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or any later version.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
+ * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
+ *
+ */
+
+package org.ironica.amatsukaze.playground
 
 import kotlinx.serialization.Serializable
-import org.ironica.amatsukaze.Direction.*
-import org.ironica.amatsukaze.Block.*
-import org.ironica.amatsukaze.Item.*
+import org.ironica.amatsukaze.playground.Direction.*
+import org.ironica.amatsukaze.playground.Block.*
+import org.ironica.amatsukaze.playground.Item.*
 
 data class Tile(var color: Color = Color.WHITE, var level: Int = 1) {
     init {
