+++ src/main/kotlin/routes/PlaygroundRoutes.kt
@@ -1,3 +1,13 @@
+/*
+ * Copyright (c) 2020-2021. kokoro-aya. All right reserved.
+ * Amatsukaze - A Playground Server implemented with ANTLR or Kotlin DSL
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or any later version.
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
+ * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
+ *
+ */
+
 package routes
 
 import io.ktor.application.Application
@@ -8,7 +18,11 @@ import io.ktor.routing.Route
 import io.ktor.routing.post
 import io.ktor.routing.route
 import io.ktor.routing.routing
-import org.ironica.amatsukaze.*
+import org.ironica.amatsukaze.bridge.AmatsukazeInterface
+import org.ironica.amatsukaze.bridge.Data
+import org.ironica.amatsukaze.bridge.convertJsonToMiscLayout
+import org.ironica.amatsukaze.bridge.convertJsonToPlayers
+import org.ironica.amatsukaze.playground.*
 import java.lang.Exception
 
 fun Route.getPlaygroundRoute(args: Pair<Boolean, Boolean>) {
